\name{measureLines-package}
\alias{measureLines-package}
\alias{measureLines}
\docType{package}
\title{
measureLines
}
\description{
Measures line segments for width, raggedness, and sharpness.
}
\details{
\tabular{ll}{
Package: \tab measureLines\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2015-05-16\cr
License: \tab GPL (>=2) | BSD\cr
}
To do \ldots
}
\author{
John Minter

Maintainer: John Minter <jrminter@gmail.com>
}
\references{
TO DO
}
\keyword{ package }
\examples{
# To do
}
